+++ npm-packages/react-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pulse-editor/react-api
 
+## 0.1.1-alpha.12
+
+### Patch Changes
+
+- Hotfix packages
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.12
+
 ## 0.1.1-alpha.11
 
 ### Patch Changes